trueblocksql: cleanup and document
